Wait longer for interface carrier to come up
In testing on physical hardware, with dhcp and static address
assignments through glean, I've encountered that the port carrier
state tends to take rougly two seconds to go from being activated
to having carrier state operational.

Changing the loop value so that we wait up to five seconds for
the interface carrier to be established before we pass on the
interface.

Change-Id: Ia524b686540596ab562fc68ca5c83e5ce38f1781