Add support for reading vlan information
This is based on the metadata format defined in the nova
metadata-service-network-info spec. While nova and neutron are still
sorting out the implementation details of when and how they will pass
this info between themselves, things like bifrost can still emit the
data and glean can be ready to recieve it.

Change-Id: I48d61963f5c45c0d3392e05e425af2331ff9b2a8