Run glean after networking service
Rather than running glean on a per-interface basis, just run
glean for all devices right before the 'networking' script runs.
This creates configuration files for all interfaces, with the info
provided by the config drive and fallback to dhcp for those not in
there.
With the current code, glean was not creating configuration files
for all interfaces probably due to a race with the network-interface
event.

Change-Id: I0885e9931111708a9861021d50324eb712a98de0