Enable network.service with systemd
glean is writing out interface scripts to
/etc/sysconfig/network-scripts ; these scripts are run by the LSB/sysv
init script /etc/init.d/network

It is quite possible this isn't enabled in systemd and on a minimal
system (no NetworkManger, cloud-config, systemd-networkd, etc),
nothing may require it, so the upshot is your system boots with no
network.

It can be confusing because the *first* time the system boots, the
glean udev rules are matched on the interface and install the
interface-scripts and then bring it up.

Enable the service when we install the interface config on redhat like
platforms.

Change-Id: Ib2b618dd975ca44e9c6b0a2c9027642ffc46b9b0