Always log debug output to syslog
Add a syslog handler and make it always output debug level log
messages.  This way, if everything zooms by on the console, at least
you've got a chance of reconstructing things from the logs/journal.

Add a streamhandler to retain the interactive output, enabled for
debug level with --debug as before.

Change-Id: Ie50acb97bbe5a2bb3c5aa11c2c327290218abd8e