Handle invalid symlinks in written files
As illustrated in I734834d03e7fdb13f9ab2e86f877b07bf4a84ff9, there are
some situations where we can have dangling symlinks on files that
glean wants to write out.

Since we're overwriting these files, it seems to make some sense to do
some best-effort detection of this.  Thus with this change, if an
output file is found to be a dangling symlink (i.e. open() returns
ELOOP), we try to unlink the file and write it again.  We only try
once and other errors are raised as usual.

The test-case is expanded to cover this path

Change-Id: I50cf89a7c419ebc4f6a5a77fca6ee03d8a3ea3bf