Fix bug in double configuring vlans on glean
There is a bug in glean that causes to configure vlans
recursively each time glean is called.
In the case of a vlan, glean is checking for sys_interfaces and
adds a new config file in the matching interface.
But if the vlan is already configured, sys_interfaces is also
listing that as a vlan, and adding the suffix again and again.

That is causing to recurrently have eth0.vlanx.cfg,
eth0.vlanx.vlanx.cfg causing network to fail on reboot or
network restart.

Change-Id: I2e930e9baa1e45b78f9362f34613945ac06e3efd