Revise systemd determination to verify systemctl presence
The base ubuntu image utilizes upstart, however it also
has a /usr/lib/systemd/system folder, which is not present
in the ubuntu-minimal images.

As a result of this, glean fails to properly work on hosts
created from the base ubuntu image. In order to address this
we should check for the presence of the systemd folder and
the systemctl utility.

Additionally, added a flag and minor logic changes to ensure
installation occurs only once.

Closes-bug: #1615673
Change-Id: I056819b4ae1e5ae1fea82fa8d34ad8cb05fe0383