Fix hostname entries in glean
glean was setting all entries for hostname to 127.0.0.1,
the same as localhost.

If was causing problem with certain use cases, for example
ansible was resolving incorrectly to localhost in boxes deployed
by glean.

As we have ip address information in network metadata, start
using that if possible, or default to 127.0.1.1 if that information
is not found. This will allow these boxes to resolve properly.

Change-Id: I400f05c84deb0fd9dbc4e1e2c8d662a82eeda55a