Mock systemlock on glean tests
systemlock is used to generate a global lock in glean.
This creates a temporary lockfile in the system, and
tests will not be creating it. Tests also fail when
running on ubuntu-trusty generated images, because that
lockfile already exists.
Mocking that call on tests seems as the most reasonable way
to proceed to fix that issue.

Change-Id: I069a9fa75e8f1828dcbab815f25d3fd2e769cb42