Add selinux context manager for writing files
Glean can be responsible for creation of a few sensitive files, so
make sure we restore the selinux context on any written files.

I have tested this with selinux enabled and disabled; the restorecon
should be a safe operation even if selinux is disabled and will ensure
the right context if you ever re-enable selinux.

Tested with centos-minimal builds and the kolla job that was
previously hitting problems with the /etc/hostname file context.

Change-Id: Ia2eba157e35b1e1b0bcdb830f98061692a198d42