Remove mock of os.stat()
This is wrong ... it's mocking os.stat() and setting everything to
octal 100777; i.e. a file with global permissions.

From what I can gather, this made it in via an old revision of
Id6cb8a62e0117f945811f3a02e6d45b769d01f31 [1] where stat was being
used to check the permissions of the file before it was opened [1].

If we want to test this, it will need to be in the open() side-effect.

[1] https://review.openstack.org/#/c/309880/5/glean/cmd.py

Change-Id: I8d9870f1cc1fa15027f876651a969ea300449aa6