Use blkid to grab config drive fs type
This simplifies our code to grab the fs type. The real motivation,
though, is we are seeing the existing code fail to return the fs type
occasionally. This appears to be another 'we run too early' bug. There
is a previous blkid command which obtains the fs type successfully,
though, and this appears to consistently work, so lets use that command
instead.

Change-Id: I7f50daa2696300686ae4315f19e23a22492dc049