Skip interfaces not in sys_interfaces
Change I30ce988fdb9459267b565b67e8bbce136fcd0249 introduced a subtle
change to these loops where interfaces that aren't in sys_interfaces
are not skipped.  It resulted in falling through and causing
KeyError's later when the non-existent MAC is looked up.

It seems the idea is that if none of the raw_macs for the interface we
are looking at (which, for a single, non-bonded, interface is just
it's MAC) are in the sys_interfaces list, we should skip
configuration.  TBH, it is still not totally clear to me what the
interaction between raw_macs and sys_interfaces is, particularly if
you do something like restrict sys_interfaces to just one interface of
a bonded interface or somethinglike that (the answer probably is --
don't do that).

This problem particularly arises under systemd were we are called for
each interface.  A test-case is added to exercise that path.

(Based on pabelanger's I049eeac6845692d20c6f18d3c2fec3fcb8e4ec34)

Change-Id: I998655808b74eb840181764ee347f2f3cad1a7e3