Ignore non permanenet addresses in udev too
Glean already ignores any interface in sysfs without a permanent address
type. It does this because interfaces with other address types are
likely managed by some other system running on the host like a neutron
agent. Add this exclusion check to the glean udev rules so that we don't
even run glean at all against them.

We have left the check in glean itself as a belts and suspenders check.

Change-Id: I893c873b047a24c0aecd6e2eb49d3f531a71f1b8