Fix SUSE based network configuration
The SUSE and openSUSE network configuration is similar to the RedHat
one but it's not identical. There are quite a few differences. There is
no network.service on SUSE. This is an alias to wicked.service and
systemd refuses to issue an 'systemctl enable <alias>'. Moreover,
the network configuration happens in ifcfg files in
/etc/sysconfg/network instead of /etc/sysconfig/network-scripts.
Furthermore, routes are being configured in /etc/sysconfig/network/ifroute-*
files instead of /etc/sysconfig/network-scripts/route-*. Finally, the
options within these files are not always the same, so we must use
whatever options make sense for each case accordingly.

Change-Id: I20bffabd333ea290d8712ec2a467f2b2d5678f3a