Always define link_type in debian interfaces
The existing code continues when an unknown link_type is found,
unfortunately we don't set link_type when the link_type is not known so
we get a python error. Just set link_type to None when we can't
determine what it is to allow the continue to function as intended.

Change-Id: Ib41650e92f24050efbc9358a6adc68421ae8a8e4