Add checks for uppercase config drive label
There are cases where the label for the config drive is all in
uppercase rather than lowercase (such as when using vfat).  The
default behaviour currently doesn't see those config drives and
fails to use them.

This patch resolves this issue by checking for config drives
with the label `config-2` and `CONFIG-2`.

Change-Id: Ibc51511f488ef41c45497795a99d05ee4e3d2d0d