Define CONFIG_DRIVE_LABEL in default case
glean.sh runs with set -eu which will cause an error if a variable is
undefined. When not using config drive and not getting a match the
variable CONFIG_DRIVE_LABEL is unset and so the -n test will fail
meaning that glean won't start and our VMs don't get network access on
boot.

Change-Id: I5fa6ce1c4a0b7218475ad5e8983df7be4ab174f4
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>