glean: Fix SUSE support for bridges and vlan configurations.
We need to check the /etc/sysconfig/network/ifcfg-* files for bridges
and vlans similar to regular interfaces. Moreover, the existing code for
checking whether the host is a SUSE distribution did not work for newer
releases (openSUSE Leap 15 is being reported as 'opensuse-leap' and
Tumbleweed is being reported as 'opensuse-tumbleweed') so we need to
refactor it a bit to actually check whether the string 'suse' exists in
the distro variable to simplify things.

Finally, the systemd service was missing the check for the SUSE network
files so we add that as well. systemd should apply a logical AND in
these conditionals so it should work fine on both Red Hat and SUSE
distros.

Change-Id: Ic5c6745dbe0077089ecaa1dd8f9b4949ac80efae