Close file in safe_open
Since this is intended to be used as a context manager, it should
close the file once we're done with it.  Also, I can't see why we
would want to reset the selinux state if an exception was raised -- it
seems there may be no file to actually do this on.  Add it to the
regular exit path.

Change-Id: I7e7dba66a1c446fa82ca3facde19f89c40e9e935