Add option to ignore config drive interfaces info
Some clouds provide incorrect network_data on the config drive because
they are doing an end round neutron to configure networking in their
cloud. Neutron has incorrect data, that ends up in the config drive and
we configure the interfaces improperly.

Address this by checking for a metadata key called
glean_ignore_interfaces. If present then we ignore any interface
configuration information in the config drive and instead fall back to
dhcp.

Change-Id: I20bc7a124891715fa2ead9e4a72956c80d63563d