Check same debian interface path everywhere
We had a special function for checking if a debian interface config file
exists but also manually constructed the path elsewhere to check if it
exists. Consolidate and drop the function since we already need the full
path in other places and once you have the full path os.path.exists is
the function you need.

Change-Id: I7a800b09699e0e86ee9df5988823b48910b57051