check networkd files
Previously we were checking Gentoo netifrc files (/etc/conf.d/net.{iname})
and not networkd files <F12>(/etc/systemd/network/{iname}.(netdev|network)).

Change to check for the networkd files.  This is not perfect as the file
name does not directly tie to the device name, but it's a start (it exists
as a Nmme or Match value within the file itself).

Change-Id: I6130102005ab61dfc25cd77c6e8abea38bccf201