Fix argv patching in unit tests
The argv patching was not actually working; it has something to do
with an interaction between the testscenarios based load_tests we use
and our mocking of argv with

 @mock.patch.object(sys, 'argv', ['./glean', '--hostname'])

What appears to happen is that tests in the same scenario groups get
the *same* reference to the sysv.argument object (i.e. it's only
created once, not once for each call).  Thus each test .append()s to
this same list, and thus we end up appending more and more arguments
to the list.  The sys.argv ends up looking like:

  stdout: {{{['./glean', '--hostname', '--distro=ubuntu', '--distro=ubuntu',
    '--distro=debian', '--distro=redhat', '--distro=redhat', '--distro=centos',
    '--distro=centos', '--distro=gentoo', '--distro=gentoo', '--distro=opensuse', ...

This can't be, Ian, how does it even work?  Well we're lucky in that
for all the existing test-cases, the final arguments appended have
overwritten the prior ones sufficiently that the test-cases work.  If
you start changing the arguments, things go wrong.

To simplify the whole thing, make main() take an argv argument that is
passed to argparse.  For normal use this comes from sys.argv, and in
unit tests we construct our own argv list for each test.

Change-Id: I090164fdd7a46d0c7a29eb6c18f80dc84cfd8469